+++ src/components/Uploader/index.js
@@ -2,22 +2,15 @@ import { h, Component } from 'preact'
 import classNames from 'classnames'
 import { isDesktop } from '~utils'
 import { camelCase } from '~utils/string'
-import { findKey } from '~utils/object'
-import { isOfMimeType } from '~utils/blob.js'
+import { validateFileTypeAndSize } from '~utils/file'
 import { trackComponentAndMode } from '../../Tracker'
 import { localised } from '../../locales'
 import theme from '../Theme/style.css'
 import style from './style.css'
-import errors from '../strings/errors'
 import CustomFileInput from '../CustomFileInput'
 import PageTitle from '../PageTitle'
 import Button from '../Button'
-
-
-const UploadError = ({ error, translate }) => {
-  const { message, instruction } = errors[error.name]
-  return <div className={style.error}>{`${translate(message)} ${translate(instruction)}`}</div>
-}
+import UploadError from './Error'
 
 const MobileUploadArea = ({ onFileSelected, children, isPoA, translate }) =>
   <div className={classNames(style.uploadArea, style.uploadAreaMobile)}>
@@ -48,57 +41,97 @@ const MobileUploadArea = ({ onFileSelected, children, isPoA, translate }) =>
     </div>
   </div>
 
-const DesktopUploadArea = ({ translate, onFileSelected, error, uploadIcon, changeFlowTo, mobileFlow }) =>
-  <div className={ style.crossDeviceInstructionsContainer }>
+const PassportMobileUploadArea = ({ nextStep, children, translate }) =>
+  <div className={classNames(style.uploadArea, style.uploadAreaMobile)}>
+    { children }
+    <div className={style.buttons}>
+      <Button
+        variants={['centered', 'primary', 'lg']}
+        onClick={nextStep}
+      >
+        {translate('capture.take_photo')}
+      </Button>
+    </div>
+  </div>
+
+const DesktopUploadArea = ({
+  translate,
+  uploadType,
+  changeFlowTo,
+  mobileFlow,
+  children
+}) => (
+  <div className={style.crossDeviceInstructionsContainer}>
     <div className={style.iconContainer}>
-      <i className={ classNames(theme.icon, style.icon, style[uploadIcon]) } />
+      <i className={classNames(theme.icon, style.icon, style[`${camelCase(uploadType)}Icon`])} />
     </div>
     <div>
-      {!mobileFlow && // Hide for mobileFlow on desktop browser as `test` Node environment has restrictedXDevice set to false
+      {!mobileFlow && ( // Hide for mobileFlow on desktop browser as `test` Node environment has restrictedXDevice set to false
         <Button
           variants={['centered', 'primary', 'lg']}
           className={ style.crossDeviceButton }
           onClick={() => changeFlowTo('crossDeviceSteps')}
         >
-          { translate('capture.switch_device') }
-        </Button>}
-      <CustomFileInput className={ style.desktopUpload } onChange={ onFileSelected }>
-        {error && <UploadError { ...{ error, translate } } />}
-        <button type="button" className={ theme.link } data-onfido-qa="uploaderButtonLink">
-          { translate('capture.upload_file') }
-        </button>
-      </CustomFileInput>
+          {translate('capture.switch_device')}
+        </Button>
+      )}
+      {children}
     </div>
   </div>
+)
 
 class Uploader extends Component {
   static defaultProps = {
-    onUpload: () => {},
-    acceptedTypes: ['jpg', 'jpeg', 'png', 'pdf'],
-    maxSize: 10000000, // The Onfido API only accepts files below 10 MB
+    onUpload: () => {}
   }
 
   setError = (name) => this.setState({ error: {name}})
 
-  findError = (file) => {
-    const { acceptedTypes, maxSize } = this.props
-    return findKey({
-      'INVALID_TYPE': file => !isOfMimeType(acceptedTypes, file),
-      'INVALID_SIZE': file => file.size > maxSize,
-    }, checkFn => checkFn(file))
-  }
-
   handleFileSelected = (file) => {
-    const error = this.findError(file)
+    const error = validateFileTypeAndSize(file)
     return error ? this.setError(error) : this.props.onUpload(file)
   }
 
-  render() {
+  renderPassportUploadIntro() {
     const {
-      title,
-      subTitle,
       changeFlowTo,
-      allowCrossDeviceFlow,
+      uploadType,
+      instructions,
+      translate,
+      mobileFlow,
+      nextStep
+    } = this.props
+    if (isDesktop) {
+      return (
+        <DesktopUploadArea
+          translate={translate}
+          uploadType={uploadType}
+          changeFlowTo={changeFlowTo}
+          mobileFlow={mobileFlow}>
+          <button
+            type="button"
+            className={theme.link}
+            data-onfido-qa="uploaderButtonLink"
+            onClick={nextStep}
+          >
+            {translate('capture.upload_file')}
+          </button>
+        </DesktopUploadArea>
+      )
+    }
+    return (
+      <PassportMobileUploadArea nextStep={nextStep} translate={translate}>
+        <div className={style.instructions}>
+          <span className={classNames(theme.icon, style.icon, style.identityIcon)} />
+          <div className={style.instructionsCopy}>{instructions}</div>
+        </div>
+      </PassportMobileUploadArea>
+    )
+  }
+
+  renderUploadArea() {
+    const {
+      changeFlowTo,
       uploadType,
       instructions,
       translate,
@@ -106,37 +139,81 @@ class Uploader extends Component {
     } = this.props
     const isPoA = uploadType === 'proof_of_address'
     const { error } = this.state
+    if (isDesktop) {
+      return (
+        <DesktopUploadArea
+          translate={translate}
+          uploadType={uploadType}
+          changeFlowTo={ changeFlowTo }
+          mobileFlow={mobileFlow}>
+          <CustomFileInput
+            className={style.desktopUpload}
+            onChange={this.handleFileSelected}
+          >
+            {error && <UploadError {...{ error, translate }} />}
+            <button
+              type="button"
+              className={theme.link}
+              data-onfido-qa="uploaderButtonLink"
+            >
+              {translate('capture.upload_file')}
+            </button>
+          </CustomFileInput>
+        </DesktopUploadArea>
+      )
+    }
     return (
-      <div className={ classNames(theme.fullHeightContainer, style.container) }>
+      <MobileUploadArea
+        onFileSelected={this.handleFileSelected}
+        translate={translate}
+        {...{ isPoA }}
+      >
+        <div className={style.instructions}>
+          <div className={classNames(style.iconContainer, {[style.poaIconContainer]: isPoA})}>
+            <span className={ classNames(
+              theme.icon,
+              style.icon,
+              style[`${camelCase(uploadType)}Icon`]
+            )} />
+          </div>
+          {error ? (
+            <UploadError {...{ error, translate }} />
+          ) : (
+            <div className={style.instructionsCopy}>{instructions}</div>
+          )}
+        </div>
+      </MobileUploadArea>
+    )
+  }
+
+  render() {
+    const {
+      title,
+      subTitle,
+      allowCrossDeviceFlow,
+      translate,
+      documentType,
+      uploadType
+    } = this.props
+    const isPassportUpload = uploadType !== "face" && documentType === 'passport'
+    return (
+      <div className={classNames(theme.fullHeightContainer, style.container)}>
         <PageTitle
           title={title}
-          subTitle={ allowCrossDeviceFlow ? translate('cross_device.switch_device.header') : subTitle } />
-        <div className={ classNames(style.uploaderWrapper, { [style.crossDeviceClient]: !allowCrossDeviceFlow }) }>
-          {isDesktop ? (
-            <DesktopUploadArea
-              onFileSelected={ this.handleFileSelected }
-              changeFlowTo={ changeFlowTo }
-              uploadIcon={ `${camelCase(uploadType)}Icon` }
-              error={error}
-              translate={translate}
-              mobileFlow={mobileFlow} />
-            ) : (
-            <MobileUploadArea
-              onFileSelected={ this.handleFileSelected }
-              translate={translate}
-              { ...{ isPoA } }
-            >
-              <div className={ style.instructions }>
-                <div className={classNames(style.iconContainer, {[style.poaIconContainer]: isPoA})}>
-                  <span className={ classNames(theme.icon, style.icon, style[`${camelCase(uploadType)}Icon`]) } />
-                </div>
-                { error ?
-                  <UploadError { ...{ error, translate } } /> :
-                  <div className={ style.instructionsCopy }>{ instructions }</div>
-                }
-              </div>
-            </MobileUploadArea>
-            )}
+          subTitle={
+            allowCrossDeviceFlow ?
+              translate('cross_device.switch_device.header') :
+              subTitle
+          }
+        />
+        <div
+          className={classNames(style.uploaderWrapper, {
+            [style.crossDeviceClient]: !allowCrossDeviceFlow,
+          })}
+        >
+          {isPassportUpload ?
+            this.renderPassportUploadIntro() :
+            this.renderUploadArea()}
         </div>
       </div>
     )
@@ -144,3 +221,8 @@ class Uploader extends Component {
 }
 
 export default trackComponentAndMode(localised(Uploader), 'file_upload', 'error')
+
+export {
+  UploadError,
+  validateFileTypeAndSize
+}