+++ test/specs/multipleBrowsersDocumentUpload.js
@@ -1,20 +1,55 @@
 import { describe, it } from '../utils/mochaw'
-import { goToPassportUploadScreen, uploadFileAndClickConfirmButton } from './scenarios/sharedFlows.js'
+import { localhostUrl } from '../config.json'
+import { goToPassportUploadScreen, uploadFileAndClickConfirmButton } from './scenarios/sharedFlows'
+
 
 const options = {
-  pageObjects: ['Confirm', 'DocumentSelector', 'Welcome', 'DocumentUpload']
+  pageObjects: [
+    'BasePage',
+    'Welcome',
+    'DocumentSelector',
+    'PassportUploadImageGuide',
+    'DocumentUpload',
+    'CrossDeviceIntro',
+    'Confirm'
+  ],
 }
 
 describe('DOCUMENT UPLOAD ON MULTIPLE BROWSERS', options, ({driver, pageObjects}) => {
-  const {welcome, documentSelector, confirm, documentUpload} = pageObjects
+
+  const {
+    basePage,
+    welcome,
+    documentSelector,
+    confirm,
+    documentUpload,
+    crossDeviceIntro,
+    passportUploadImageGuide
+  } = pageObjects
+
+  const copy = basePage.copy()
 
   it('should upload document with JPG', async () => {
     goToPassportUploadScreen(driver, welcome, documentSelector)
-    uploadFileAndClickConfirmButton(documentUpload, confirm, 'passport.jpg')
+    documentUpload.clickUploadButton()
+    driver.sleep(500)  // add delay to account for IE11 taking a while to transition to this screen on Browserstack
+    uploadFileAndClickConfirmButton(passportUploadImageGuide, confirm, 'passport.jpg')
   })
 
   it('should upload document with PDF', async () => {
-    goToPassportUploadScreen(driver, welcome, documentSelector)
+    driver.get(localhostUrl)
+    welcome.continueToNextStep()
+    documentSelector.clickOnIdentityCardIcon()
     uploadFileAndClickConfirmButton(documentUpload, confirm, 'national_identity_card.pdf')
+  })
+
+  it('should show cross device intro screen if camera not detected and uploadFallback disabled', async () => {
+    goToPassportUploadScreen(driver, welcome, documentSelector, `?uploadFallback=false`)
+    documentUpload.clickUploadButton()
+    driver.sleep(500)  // add delay to account for IE11 taking a while to transition to this screen on Browserstack
+    uploadFileAndClickConfirmButton(passportUploadImageGuide, confirm, 'passport.jpg')
+    crossDeviceIntro.verifyTitle(copy)
+    crossDeviceIntro.verifySubTitle(copy)
+    crossDeviceIntro.verifyMessages(copy)
   })
 })